[🛠️Fix] Player
- 플레이어 스크립트 수정
