+++ 2DRoguelike/Assets/Resources/Prefabs/Models/Player.prefab
@@ -1,5 +1,36 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &2052860844766876189
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 4363092316223236387}
+  m_Layer: 0
+  m_Name: ProjectileSpawnPoint
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4363092316223236387
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2052860844766876189}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 6094545270946103118}
+  m_RootOrder: -1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &6720210768200686873
 GameObject:
   m_ObjectHideFlags: 0
@@ -12,9 +43,7 @@ GameObject:
   - component: {fileID: 6286260149814370698}
   - component: {fileID: 9089910657805106896}
   - component: {fileID: 5392859993858710398}
-  - component: {fileID: 6467323381560477376}
   - component: {fileID: 6700628932903463887}
-  - component: {fileID: 8797934997506717847}
   - component: {fileID: 3848652982911187825}
   m_Layer: 0
   m_Name: Player
@@ -37,7 +66,7 @@ Transform:
   m_Children:
   - {fileID: 6094545270946103118}
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!50 &6286260149814370698
 Rigidbody2D:
@@ -109,7 +138,7 @@ BoxCollider2D:
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 0.25, y: 0.3}
+  m_Size: {x: 1, y: 1.4}
   m_EdgeRadius: 0
 --- !u!114 &5392859993858710398
 MonoBehaviour:
@@ -141,18 +170,6 @@ MonoBehaviour:
   m_DefaultActionMap: Player
   m_SplitScreenIndex: -1
   m_Camera: {fileID: 0}
---- !u!114 &6467323381560477376
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6720210768200686873}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5d9e07b17025f944fb96061814db84f0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &6700628932903463887
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -165,19 +182,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 25bd8751de6e4364da13a783c9208162, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &8797934997506717847
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6720210768200686873}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 950df1cc49b1e0d43820508b4bea1763, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
   characterRenderer: {fileID: 1899615088405746631}
+  projectileSpawnPosition: {fileID: 4363092316223236387}
+  testPrefab: {fileID: 2645124887727915881, guid: 92fbd77f7e7fa90449a93763c94c2c7d, type: 3}
 --- !u!114 &3848652982911187825
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -216,9 +223,10 @@ Transform:
   m_GameObject: {fileID: 8511564257792373920}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 0.4, y: 0.4, z: 1}
+  m_LocalScale: {x: 2, y: 2, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children: []
+  m_Children:
+  - {fileID: 4363092316223236387}
   m_Father: {fileID: 1139828893238760913}
   m_RootOrder: -1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
