--- 2DRoguelike/Assets/Scripts/Player/AimRotation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 950df1cc49b1e0d43820508b4bea1763
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
