+++ 2DRoguelike/Assets/Scripts/Player/CharacterController.cs
@@ -8,27 +8,12 @@ public class CharacterController : MonoBehaviour
     public event Action<Vector2> OnMoveEvent;
     public event Action<Vector2> OnLookEvent;
     public event Action<Vector2> OnAttackEvent;
-    private float _timeSinceLastAttack = float.MaxValue;
-    protected bool IsAttacking { get; set; }
-
+    
     protected virtual void Update()
     {
-        AttackDelay();
+
     }
 
-    private void AttackDelay()
-    {
-        if (_timeSinceLastAttack <= 0.5f)    
-        {
-            _timeSinceLastAttack += Time.deltaTime;
-        }
-
-        if (IsAttacking && _timeSinceLastAttack > 0.5f)
-        {
-            _timeSinceLastAttack = 0;
-            CallAttackEvent(Vector2.zero);
-        }
-    }
     public void CallMoveEvent(Vector2 direction)
     {
         OnMoveEvent?.Invoke(direction);
