--- 2DRoguelike/Assets/Scripts/Player/Movement.cs
@@ -1,38 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Movement : MonoBehaviour
-{
-    private CharacterController _controller;
-    private Vector2 _movementDirection = Vector2.zero;
-    private Rigidbody2D _rigidbody;
-
-    private void Awake()
-    {
-        _controller = GetComponent<CharacterController>();
-        _rigidbody = GetComponent<Rigidbody2D>();
-    }
-
-    private void Start()
-    {
-        _controller.OnMoveEvent += Move;
-    }
-
-    private void FixedUpdate()
-    {
-        ApplyMovment(_movementDirection);
-    }
-
-    private void Move(Vector2 direction)
-    {
-        _movementDirection = direction;
-    }
-
-    private void ApplyMovment(Vector2 direction)
-    {
-        direction = direction * 5;  //플레이어 speed
-
-        _rigidbody.velocity = direction;
-    }
-}
