--- 2DRoguelike/Assets/Scripts/Player/Movement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5d9e07b17025f944fb96061814db84f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
