+++ 2DRoguelike/Assets/Scripts/Player/PlayerInputController.cs
@@ -1,33 +1,164 @@
 using System.Collections;
 using System.Collections.Generic;
+using System.Security.Claims;
 using UnityEngine;
 using UnityEngine.InputSystem;
+using UnityEngine.InputSystem.XR;
 
 public class PlayerInputController : CharacterController
 {
+    #region Fileds
+
+    [SerializeField] private SpriteRenderer characterRenderer;
+    [SerializeField] private Transform projectileSpawnPosition;
+
+    public GameObject testPrefab;
+
+    #endregion
+
+    #region Properties
+
     private Camera _camera;
+    private Vector2 _movementDirection = Vector2.zero;
+    private Vector2 _aimDirection = Vector2.right;
+    private Rigidbody2D _rigidbody;
+    private Player _player;
+    private float _timeSinceLastAttack = float.MaxValue;
+    protected bool IsAttacking { get; set; }
+
+    #endregion
+
     private void Awake()
     {
         _camera = Camera.main;
+        _rigidbody = GetComponent<Rigidbody2D>();
+        _player = GetComponent<Player>();
     }
 
+    private void Start()
+    {
+        OnMoveEvent += Move;
+        OnLookEvent += OnAim;
+        OnAttackEvent += OnShoot;
+    }
+
+    protected override void Update()
+    {
+        base.Update(); 
+        AttackDelay(); 
+    }
+
+   
+
+    #region Move
     public void OnMove(InputValue value)
     {
         Vector2 moveInput = value.Get<Vector2>().normalized;
         CallMoveEvent(moveInput);
     }
 
+    private void FixedUpdate()
+    {
+        ApplyMovment(_movementDirection);
+    }
+
+    private void Move(Vector2 direction)
+    {
+        _movementDirection = direction;
+    }
+
+    private void ApplyMovment(Vector2 direction)
+    {
+        direction = direction * 5;  //플레이어 speed
+
+        _rigidbody.velocity = direction;
+    }
+    #endregion
+
+    #region Look
     public void OnLook(InputValue value)
     {
         Vector2 newAim = value.Get<Vector2>();
         CallLookEvent(newAim);
     }
 
+    public void OnAim(Vector2 newAimDirection)
+    {
+        _aimDirection = newAimDirection.normalized;
+        RotatePlayer(newAimDirection);
+    }
+
+    private void RotatePlayer(Vector2 direction)
+    {
+        if (direction.x > 0)
+        {
+            // 오른쪽 방향일 때
+            characterRenderer.flipX = false;
+        }
+        else if (direction.x < 0)
+        {
+            // 왼쪽 방향일 때
+            characterRenderer.flipX = true;
+        }
+
+        // 방향에 따라 스프라이트를 수직으로 뒤집거나 되돌립니다.
+        if (direction.y > 0)
+        {
+            // 위쪽 방향일 때
+            characterRenderer.flipY = false;
+        }
+        else if (direction.y < 0)
+        {
+            // 아래쪽 방향일 때
+            characterRenderer.flipY = true;
+        }
+    }
+    #endregion
+
+    #region Attack
     public void OnFire(InputValue value)
     {
         Vector2 fireValue = value.Get<Vector2>();
-        CallAttackEvent(fireValue);
+        IsAttacking = fireValue.magnitude > 0f;
     }
+
+    private void OnShoot(Vector2 direction)
+    {
+        CreateProjectile();
+    }
+
+    private void CreateProjectile()
+    {
+        // 발사 방향으로 회전된 각도 계산
+        float angle = Mathf.Atan2(_aimDirection.y, _aimDirection.x) * Mathf.Rad2Deg;
+        Quaternion rotation = Quaternion.AngleAxis(angle, Vector3.forward);
+
+        // 발사체 생성
+        GameObject projectile = Instantiate(testPrefab, projectileSpawnPosition.position, rotation);
+
+        Rigidbody2D projectileRb = projectile.GetComponent<Rigidbody2D>();
+        if (projectileRb == null)
+        {
+            projectileRb = projectile.AddComponent<Rigidbody2D>();
+        }
+        projectileRb.velocity = _aimDirection * _player.ShotSpeed;
+    }
+    
+    private void AttackDelay()
+    {
+        if (_timeSinceLastAttack <= _player.AttackSpeed)
+        {
+            _timeSinceLastAttack += Time.deltaTime;
+        }
+
+        if (IsAttacking && _timeSinceLastAttack > _player.AttackSpeed)
+        {
+            _timeSinceLastAttack = 0;
+            CallAttackEvent(Vector2.zero);
+        }
+    }
+    #endregion
+
 
     public void OnBoom(InputValue value)
     {
