--- 2DRoguelike/Assets/Scripts/Player/Shooting.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a24df707b8ad8b54dbefc03224ed1e2d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
